+++ bin/impaktfull_cli.dart
@@ -1,3 +1,4 @@
 import 'package:impaktfull_cli/src/impaktfull_cli.dart';
 
-Future<void> main(List<String> arguments) => ImpaktfullCli(arguments: arguments).runCli();
+Future<void> main(List<String> arguments) =>
+    ImpaktfullCli(arguments: arguments).runCli();